BAU: remove boilerplate in test scoped *ClientRule classes which extend DropwizardClientRule; we can use DropwizardClientRule directly, and supply a Resource
