--- proxy-node-gateway/src/test/java/uk/gov/ida/notification/apprule/rules/EidasSamlParserClientRule.java
@@ -1,9 +1,0 @@
-package uk.gov.ida.notification.apprule.rules;
-
-import io.dropwizard.testing.junit.DropwizardClientRule;
-
-public class EidasSamlParserClientRule<T> extends DropwizardClientRule {
-    public EidasSamlParserClientRule(T resource) {
-        super(resource);
-    }
-}
