--- proxy-node-gateway/src/test/java/uk/gov/ida/notification/apprule/rules/TranslatorClientRule.java
@@ -1,7 +1,0 @@
-package uk.gov.ida.notification.apprule.rules;
-
-import io.dropwizard.testing.junit.DropwizardClientRule;
-
-public class TranslatorClientRule<T> extends DropwizardClientRule {
-    public TranslatorClientRule(T resource) { super(resource); }
-}
