--- proxy-node-gateway/src/test/java/uk/gov/ida/notification/apprule/rules/VerifyServiceProviderClientRule.java
@@ -1,9 +1,0 @@
-package uk.gov.ida.notification.apprule.rules;
-
-import io.dropwizard.testing.junit.DropwizardClientRule;
-
-public class VerifyServiceProviderClientRule<T> extends DropwizardClientRule {
-    public VerifyServiceProviderClientRule(T resource) {
-        super(resource);
-    }
-}
