--- stub-connector/src/test/java/uk/gov/ida/notification/apprule/rules/ProxyNodeClientRule.java
@@ -1,9 +1,0 @@
-package uk.gov.ida.notification.apprule.rules;
-
-import io.dropwizard.testing.junit.DropwizardClientRule;
-
-public class ProxyNodeClientRule extends DropwizardClientRule {
-    public ProxyNodeClientRule() {
-        super(new TestProxyNodeResource());
-    }
-}
